REJECT This push, I just want to get rid of the commit files in master branch

+++ src/main/java/com/prison/project/model/PrisonCapacity.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class PrisonCapacity {
 
-    private final Integer capacity = 10;
+    private final Long capacity = 10L;
